@@ -11,8 +11,8 @@
     "url": "https://github.com/ngxs-labs/async-storage-plugin/issues"
   },
   "peerDependencies": {
-    "@angular/core": ">=6.1.0 <10.0.0",
-    "@ngxs/store": ">=3.6.0",
-    "rxjs": ">=6.0.0"
+    "@angular/core": "^16.2.12",
+    "@ngxs/store": "^3.8.2",
+    "rxjs": "^7.4.0"
   }
 }
