@@ -1,17 +1,8 @@
-import {
-  NgModule,
-  ModuleWithProviders,
-  InjectionToken,
-  TypeProvider
-} from '@angular/core';
+import { InjectionToken, ModuleWithProviders, NgModule, TypeProvider } from '@angular/core';
 import { NGXS_PLUGINS } from '@ngxs/store';
 
 import { NgxsAsyncStoragePlugin } from './async-storage.plugin';
-import {
-  NgxsStoragePluginOptions,
-  NGXS_STORAGE_PLUGIN_OPTIONS,
-  STORAGE_ENGINE
-} from './symbols';
+import { NGXS_STORAGE_PLUGIN_OPTIONS, NgxsStoragePluginOptions, STORAGE_ENGINE } from './symbols';
 import { storageOptionsFactory } from './internals';
 
 export const USER_OPTIONS = new InjectionToken('USER_OPTIONS');
@@ -21,7 +12,7 @@ export class NgxsAsyncStoragePluginModule {
   static forRoot(
     engine: TypeProvider,
     options?: NgxsStoragePluginOptions
-  ): ModuleWithProviders {
+  ): ModuleWithProviders<NgxsAsyncStoragePluginModule> {
     return {
       ngModule: NgxsAsyncStoragePluginModule,
       providers: [
